src/requirements.txt
Authlib==1.3.2
asyncpg==0.29.0
alembic==1.13.2
argon2-cffi==23.1.0
fastapi==0.114.2
uvicorn==0.30.6
pydantic==2.9.1
pydantic-settings==2.5.2
sqlalchemy==2.0.34
starlette==0.38.5
dishka==1.4.0
PyJWT==2.10.0
redis==5.2.0
ruff==0.8.0
coloredlogs==15.0.1
